+++ packages/suspense-website/src/routes/api/createCache.tsx
@@ -132,7 +132,7 @@ export default function Route() {
       <Block>
         <SubHeading title="Observing status" />
         <p>
-          A value's <em>status</em> ("not-started", "pending", "resolved", or
+          A value's <em>status</em> ("not-found", "pending", "resolved", or
           "rejected") can be queried as well using <code>cache.getStatus</code>–
           although the recommended way to subscribe to this value is using the{" "}
           <code>